Cast double to float

+++ cube/linmath.h
@@ -601,7 +601,7 @@ LINMATH_H_FUNC void mat4x4_arcball(mat4x4 R, mat4x4 const M, vec2 const _a, vec2
 	vec3 c_;
 	vec3_mul_cross(c_, a_, b_);
 
-	float const angle = acos(vec3_mul_inner(a_, b_)) * s;
+	float const angle = acosf(vec3_mul_inner(a_, b_)) * s;
 	mat4x4_rotate(R, M, c_[0], c_[1], c_[2], angle);
 }
